Working on file download for Firefox
Right now file download works in Firefox but not in Chrome.
This is because blobs can't be passed from background.js to dashboard.js
through "chrome.runtime.sendMessage" in Chrome, this works in Firefox
however.

Possible fix would be to use the previous code for Chrome and the new
for Firefox?